✨ feat(ping): add advanced latency statistics and increase log visibility
add standard deviation and 99th percentile latency to the summary table for a more detailed view of connection stability
increase the number of displayed log entries from 15 to 70 to improve real-time diagnostics
