+++ src/ZtrConnectionTester.Console/Commands/PingPongPair/PingCommand/LogEntry.cs
@@ -1,0 +1,5 @@
+﻿using System;
+
+namespace ZtrConnectionTester.Console.Commands.PingPongPair.SerialPing;
+
+public record LogEntry(DateTime Timestamp, string Message, LogLevel Level, string? Source = null);
