+++ src/ZtrConnectionTester.Console/Commands/PingPongPair/PingCommand/LogLevel.cs
@@ -1,0 +1,3 @@
+﻿namespace ZtrConnectionTester.Console.Commands.PingPongPair.SerialPing;
+
+public enum LogLevel { Detail, Info, Warning, Error }
