+++ src/ZtrConnectionTester.Console/Commands/PingPongPair/PingCommand/SerialPingSettings.cs
@@ -1,0 +1,8 @@
+﻿using ZtrConnectionTester.Console.Commands.PingPongPair.Commons;
+
+namespace ZtrConnectionTester.Console.Commands.PingPongPair.SerialPing;
+
+public class SerialPingSettings(ISerialPortEnumerator portEnumerator) : BaseSerialPingSettings(portEnumerator)
+{
+
+}
